+++ was_bug12/was_bug12_jar/src/main/java/at/struct/wasbugs/wasbug12/be/CacheUsingService.java
@@ -1,4 +1,4 @@
-package at.struct.wasbugs.wasbug12;
+package at.struct.wasbugs.wasbug12.be;
 
 import javax.enterprise.context.ApplicationScoped;
